moved TypeIdMap to bevy_utils
