+++ crates/bevy_asset/src/server/info.rs
@@ -4,9 +4,9 @@ use crate::{
     Handle, InternalAssetEvent, LoadState, RecursiveDependencyLoadState, StrongHandle,
     UntypedAssetId, UntypedHandle,
 };
-use bevy_ecs::{world::World, TypeIdMap};
+use bevy_ecs::world::World;
 use bevy_log::warn;
-use bevy_utils::{Entry, HashMap, HashSet};
+use bevy_utils::{Entry, HashMap, HashSet, TypeIdMap};
 use crossbeam_channel::Sender;
 use std::{
     any::TypeId,
@@ -205,7 +205,7 @@ impl AssetInfos {
             .ok_or(GetOrCreateHandleInternalError::HandleMissingButTypeIdNotSpecified)?;
 
         match handles.entry(type_id) {
-            std::collections::hash_map::Entry::Occupied(entry) => {
+            bevy_utils::hashbrown::hash_map::Entry::Occupied(entry) => {
                 let id = *entry.get();
                 // if there is a path_to_id entry, info always exists
                 let info = self.infos.get_mut(&id).unwrap();
@@ -246,7 +246,7 @@ impl AssetInfos {
                 }
             }
             // The entry does not exist, so this is a "fresh" asset load. We must create a new handle
-            std::collections::hash_map::Entry::Vacant(entry) => {
+            bevy_utils::hashbrown::hash_map::Entry::Vacant(entry) => {
                 let should_load = match loading_mode {
                     HandleLoadingMode::NotLoading => false,
                     HandleLoadingMode::Request | HandleLoadingMode::Force => true,
